HACK:
Usually in most 3D models all verices have bones that infuence them
and usually they sum up to one.

However... from Sketchfab I found a mesh that had a few vertices that did not
have any bones infuencing them.
In that case, I guess, the correct solution would be to use the transformation
of the node where the mesh is attached for that vertex.
This however would introduce more special cases in the code to be handled.
(maybe this is what I should do).

In order to avoid that, if the mesh has any bones,
for each vertex that is not infuenced by anything,
we find the closest bone to it and add 100% weight to it
that way the vertex is going to get transformed.
This is probably produce unwanted motion but it should be minimal.